feat(install): add sh and shell recipe executors
