+++ internal/install/execution/go_task_recipe_executor.go
@@ -2,70 +2,39 @@ package execution
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
-	"strconv"
 	"strings"
 
-	survey "github.com/AlecAivazis/survey/v2"
-	"github.com/AlecAivazis/survey/v2/terminal"
 	"github.com/go-task/task/v3"
 	taskargs "github.com/go-task/task/v3/args"
 	"github.com/go-task/task/v3/taskfile"
 	log "github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 
-	"github.com/newrelic/newrelic-cli/internal/credentials"
 	"github.com/newrelic/newrelic-cli/internal/install/types"
 )
 
 // GoTaskRecipeExecutor is an implementation of the recipeExecutor interface that
 // uses the go-task module to execute the steps defined in each recipe.
-type GoTaskRecipeExecutor struct{}
+type GoTaskRecipeExecutor struct {
+	Stdin  io.Reader
+	Stdout io.Writer
+	Stderr io.Writer
+}
 
 // NewGoTaskRecipeExecutor returns a new instance of GoTaskRecipeExecutor.
 func NewGoTaskRecipeExecutor() *GoTaskRecipeExecutor {
-	return &GoTaskRecipeExecutor{}
+	return &GoTaskRecipeExecutor{
+		Stdin:  os.Stdin,
+		Stdout: os.Stdout,
+		Stderr: os.Stderr,
+	}
 }
 
-func (re *GoTaskRecipeExecutor) Prepare(ctx context.Context, m types.DiscoveryManifest, r types.OpenInstallationRecipe, assumeYes bool, licenseKey string) (types.RecipeVars, error) {
-	log.WithFields(log.Fields{
-		"name": r.Name,
-	}).Debug("preparing recipe")
-
-	vars := types.RecipeVars{}
-
-	results := []types.RecipeVars{}
-
-	systemInfoResult := varsFromSystemInfo(m)
-
-	profileResult, err := varsFromProfile(licenseKey)
-	if err != nil {
-		return types.RecipeVars{}, err
-	}
-
-	inputVarsResult, err := varsFromInput(r.InputVars, assumeYes)
-	if err != nil {
-		return types.RecipeVars{}, err
-	}
-
-	results = append(results, systemInfoResult)
-	results = append(results, profileResult)
-	results = append(results, types.RecipeVariables)
-	results = append(results, inputVarsResult)
-
-	for _, result := range results {
-		for k, v := range result {
-			vars[k] = v
-		}
-	}
-
-	return vars, nil
-}
-
-func (re *GoTaskRecipeExecutor) Execute(ctx context.Context, m types.DiscoveryManifest, r types.OpenInstallationRecipe, recipeVars types.RecipeVars) error {
+func (re *GoTaskRecipeExecutor) Execute(ctx context.Context, r types.OpenInstallationRecipe, recipeVars types.RecipeVars) error {
 	log.Debugf("executing recipe %s", r.Name)
 
 	out := []byte(r.Install)
@@ -82,14 +51,14 @@ func (re *GoTaskRecipeExecutor) Execute(ctx context.Context, m types.DiscoveryMa
 		return err
 	}
 
-	stdoutCapture := NewLineCaptureBuffer(os.Stdout)
-	stderrCapture := NewLineCaptureBuffer(os.Stderr)
+	stdoutCapture := NewLineCaptureBuffer(re.Stdout)
+	stderrCapture := NewLineCaptureBuffer(re.Stderr)
 
 	e := task.Executor{
 		Entrypoint: file.Name(),
 		Stderr:     stderrCapture,
 		Stdout:     stdoutCapture,
-		Stdin:      os.Stdin,
+		Stdin:      re.Stdin,
 	}
 
 	if err = e.Setup(); err != nil {
@@ -138,116 +107,4 @@ func (re *GoTaskRecipeExecutor) Execute(ctx context.Context, m types.DiscoveryMa
 	}
 
 	return nil
-}
-
-func varsFromProfile(licenseKey string) (types.RecipeVars, error) {
-	defaultProfile := credentials.DefaultProfile()
-	if licenseKey == "" {
-		return types.RecipeVars{}, errors.New("license key not found")
-	}
-
-	vars := make(types.RecipeVars)
-
-	vars["NEW_RELIC_LICENSE_KEY"] = licenseKey
-	vars["NEW_RELIC_ACCOUNT_ID"] = strconv.Itoa(defaultProfile.AccountID)
-	vars["NEW_RELIC_API_KEY"] = defaultProfile.APIKey
-	vars["NEW_RELIC_REGION"] = defaultProfile.Region
-
-	return vars, nil
-}
-
-func varsFromSystemInfo(m types.DiscoveryManifest) types.RecipeVars {
-	vars := make(types.RecipeVars)
-
-	vars["HOSTNAME"] = m.Hostname
-	vars["OS"] = m.OS
-	vars["PLATFORM"] = m.Platform
-	vars["PLATFORM_FAMILY"] = m.PlatformFamily
-	vars["PLATFORM_VERSION"] = m.PlatformVersion
-	vars["KERNEL_ARCH"] = m.KernelArch
-	vars["KERNEL_VERSION"] = m.KernelVersion
-
-	return vars
-}
-
-func varsFromInput(inputVars []types.OpenInstallationRecipeInputVariable, assumeYes bool) (types.RecipeVars, error) {
-	vars := make(types.RecipeVars)
-
-	vars["NEW_RELIC_ASSUME_YES"] = fmt.Sprintf("%t", assumeYes)
-
-	for _, envConfig := range inputVars {
-		var err error
-		envValue := os.Getenv(envConfig.Name)
-
-		if envValue != "" {
-			vars[envConfig.Name] = envValue
-			continue
-		}
-
-		if assumeYes {
-			if envConfig.Default == "" {
-				return types.RecipeVars{}, fmt.Errorf("no default value for environment variable %s and none provided", envConfig.Name)
-			}
-
-			log.WithFields(log.Fields{
-				"name":    envConfig.Name,
-				"default": envConfig.Default,
-			}).Debug("required env var not found, using default")
-
-			envValue = envConfig.Default
-		} else {
-			log.WithFields(log.Fields{
-				"name": envConfig.Name,
-			}).Debug("required environment variable not found")
-
-			envValue, err = varFromPrompt(envConfig)
-			if err != nil {
-				if err == terminal.InterruptErr {
-					return types.RecipeVars{}, types.ErrInterrupt
-				}
-
-				return types.RecipeVars{}, fmt.Errorf("prompt failed: %s", err)
-			}
-		}
-
-		vars[envConfig.Name] = envValue
-	}
-
-	return vars, nil
-}
-
-func varFromPrompt(envConfig types.OpenInstallationRecipeInputVariable) (string, error) {
-	msg := fmt.Sprintf("value for %s required", envConfig.Name)
-
-	if envConfig.Prompt != "" {
-		msg = envConfig.Prompt
-	}
-
-	value := ""
-	var prompt survey.Prompt
-
-	if envConfig.Secret {
-		prompt = &survey.Password{
-			Message: msg,
-		}
-	} else {
-		defaultValue := ""
-
-		if envConfig.Default != "" {
-			defaultValue = envConfig.Default
-		}
-
-		prompt = &survey.Input{
-			Message: msg,
-			Default: defaultValue,
-		}
-	}
-
-	err := survey.AskOne(prompt, &value)
-	if err != nil {
-		return "", err
-	}
-
-	return value, nil
-
 }
