+++ internal/install/execution/recipe_executor.go
@@ -9,6 +9,5 @@ import (
 // RecipeExecutor is responsible for execution of the task steps defined in a
 // recipe.
 type RecipeExecutor interface {
-	Prepare(context.Context, types.DiscoveryManifest, types.OpenInstallationRecipe, bool, string) (types.RecipeVars, error)
-	Execute(context.Context, types.DiscoveryManifest, types.OpenInstallationRecipe, types.RecipeVars) error
+	Execute(context.Context, types.OpenInstallationRecipe, types.RecipeVars) error
 }
