+++ internal/install/recipe_installer_test.go
@@ -42,6 +42,7 @@ var (
 	pi              = ux.NewMockProgressIndicator()
 	lkf             = NewMockLicenseKeyFetcher()
 	cv              = diagnose.NewMockConfigValidator()
+	rvp             = execution.NewConcreteRecipeVarProvider()
 )
 
 func TestInstall(t *testing.T) {
@@ -58,7 +59,7 @@ func TestNewRecipeInstaller_InstallerContextFields(t *testing.T) {
 		SkipApm:            true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 
 	require.True(t, reflect.DeepEqual(ic, i.InstallerContext))
 }
@@ -67,7 +68,7 @@ func TestShouldGetRecipeFromURL(t *testing.T) {
 	ic := InstallerContext{}
 	ff = recipes.NewMockRecipeFileFetcher()
 	ff.FetchRecipeFileFunc = fetchRecipeFileFunc
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 
 	recipe, err := i.recipeFromPath("http://recipe/URL")
 	require.NoError(t, err)
@@ -79,7 +80,7 @@ func TestShouldGetRecipeFromFile(t *testing.T) {
 	ic := InstallerContext{}
 	ff = recipes.NewMockRecipeFileFetcher()
 	ff.LoadRecipeFileFunc = loadRecipeFileFunc
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 
 	recipe, err := i.recipeFromPath("file.txt")
 	require.NoError(t, err)
@@ -94,7 +95,7 @@ func TestInstall_Basic(t *testing.T) {
 		{Name: types.InfraAgentRecipeName},
 		{Name: types.LoggingRecipeName},
 	}
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, f.FetchRecipeNameCount[types.InfraAgentRecipeName], 1)
@@ -114,7 +115,7 @@ func TestInstall_DiscoveryComplete(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 
 	err := i.Install()
 	require.NoError(t, err)
@@ -137,7 +138,7 @@ func TestInstall_FailsOnInvalidOs(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, discover, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, discover, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 
 	err := i.Install()
 	require.Error(t, err)
@@ -165,7 +166,7 @@ func TestInstall_RecipesAvailable(t *testing.T) {
 		},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipesAvailableCallCount)
@@ -207,7 +208,7 @@ func TestInstall_RecipeInstalled(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f2, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f2, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 3, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -246,7 +247,7 @@ func TestInstall_RecipeFailed(t *testing.T) {
 	v = validation.NewMockRecipeValidator()
 	v.ValidateErr = errors.New("validationErr")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 1, v.ValidateCallCount)
@@ -271,7 +272,7 @@ func TestInstall_InstallComplete(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -290,7 +291,7 @@ func TestInstall_InstallCanceled(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -319,7 +320,7 @@ func TestInstall_InstallCompleteError(t *testing.T) {
 	v = validation.NewMockRecipeValidator()
 	v.ValidateErr = errors.New("test error")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.Error(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -360,7 +361,7 @@ func TestInstall_InstallCompleteError_guidedRecipeFail(t *testing.T) {
 		errors.New("testing error"),
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).InstallCompleteCallCount)
@@ -396,7 +397,7 @@ func TestInstall_RecipeSkipped(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -434,7 +435,7 @@ func TestInstall_RecipeSkippedApm(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -472,7 +473,7 @@ func TestInstall_RecipeSkippedApmAnyKeyword(t *testing.T) {
 		PromptMultiSelectAll: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -509,7 +510,7 @@ func TestInstall_RecipeSkipped_SkipAll(t *testing.T) {
 		PromptMultiSelectVal: []string{},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -545,7 +546,7 @@ func TestInstall_RecipeSkipped_MultiSelect(t *testing.T) {
 		PromptMultiSelectVal: []string{testRecipeName},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -608,7 +609,7 @@ func TestInstall_RecipeRecommended(t *testing.T) {
 		PromptMultiSelectVal: []string{testRecipeName},
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -650,7 +651,7 @@ func TestInstall_RecipeSkipped_AssumeYes(t *testing.T) {
 		PromptYesNoVal: true,
 	}
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).RecipeSkippedCallCount)
@@ -676,7 +677,7 @@ func TestInstall_TargetedInstall_InstallsInfraAgent(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -706,7 +707,7 @@ func TestInstall_TargetedInstall_InstallsInfraAgentDependency(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -731,7 +732,7 @@ func TestInstall_TargetedInstallInfraAgent_NoInfraAgentDuplicate(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -757,7 +758,7 @@ func TestInstall_TargetedInstall_SkipInfra(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 0, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -788,7 +789,7 @@ func TestInstall_TargetedInstall_SkipInfraDependency(t *testing.T) {
 
 	v = validation.NewMockRecipeValidator()
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 1, statusReporters[0].(*execution.MockStatusReporter).RecipeInstalledCallCount)
@@ -835,7 +836,7 @@ func TestInstall_GuidReport(t *testing.T) {
 	// Test for NEW_RELIC_CLI_VERSION
 	os.Setenv("NEW_RELIC_CLI_VERSION", "testversion0.0.1")
 
-	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv}
+	i := RecipeInstaller{ic, d, l, mv, f, e, v, ff, status, p, pi, lkf, cv, rvp}
 	err := i.Install()
 	require.NoError(t, err)
 	require.Equal(t, 2, v.ValidateCallCount)
